Stop using the URL fragment to find the first unread post on a page
The fragment takes into account deleted posts, but Awful does not get to see deleted posts, so we would unhelpfully hide everything on the page and then crash if you tapped "Previous Posts". If the fragment is absent (e.g. by visiting via bookmarks or a direct link with no fragment) then Awful did not get erroneous info and all was well.

Fixes #673.