Bump minimum SDK back down to iOS 9.0
Turns out we have a few people stuck on past iOS versions, and the Forums markup changed in a way that makes past Awful versions largely unusable. In the past I would bump the minimum SDK up in the name of easier code maintenance, but in practice I would not say we realized this supposed benefit. So let's try going the other way and extending support back as far as sensible.

Xcode 11.3, the last version of Xcode to run on macOS 10.14 and thus the last version to support the iOS 9 simulator, does not support the throwing variants of XCTest setUp/tearDown methods. Also, the stickers extension must have a deployment target of 10.0 or Awful fails to launch.

Alternate app icons require iOS 10.3, so we'll hide that from Settings when appropriate.